Implemented usernames
Usernames are key to enabling the social experience of sharing karaoke repertoire. We add a username field to the Devise user table, update the User model to acknowledge and validate the new field, update the Devise views, and enable the friendly_id gem to allow for easy slugging (along with simplified routes).